# Threaded MultipackDistributedDataloader with prefetched samples