--- a/src/axolotl/utils/dataloader.py
+++ b/src/axolotl/utils/dataloader.py
@@ -3,14 +3,14 @@ import hashlib
 import itertools
 import logging
 import math
+import time
+from collections import deque
+from threading import Thread
 from typing import Any, Callable, List, Union
 
 import numba
 import numpy as np
 from torch.utils.data import DistributedSampler, Sampler
-import time
-from collections import deque
-from threading import Thread
 
 LOG = logging.getLogger("axolotl.utils.dataloader")